+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `step_collapse_cart()` can pool a predictor's factor levels using a tree-based method. 
 
+* `step_collapse_stringdist()` can pool a predictor's factor levels using string distances.
+
 # embed 0.2.0
 
 * `step_embed()` now correctly defaults to have a random id with the word "embed". (#102)